Supporting Old and New Format
Changes were made in export_IFSM_Weather.m to support both new format
and old format. Old Format did not have mean temperature and wind speed.
Also the program now make sure that there are only 365 days in a year.
For leap year, the last record is dropped to make it down to 366 days.
That corresponds to Dec 31. It was decided to do that to avoid one day
shift within the growing season.